Auto-tune at 0.5 MHz steps; menu generated to match
Scan candidates are now 8.0-20.0 in 0.5 MHz steps (25) so a clean band that's
centered on a half-MHz on some WiFi channel still gets found (the integer grid
could straddle a band with dirty notches either side). Scan/persist/status all
carry one-decimal MHz; the XCLK menu is generated in JS at 0.5 steps and syncs
to the device's actual value.

Scanned both: S3 -> 14.5 MHz (clean band); AI-Thinker found nothing clean
(marginal link, all <=4 fps) so it stays at safe 8.

+++ src/Camera.h
@@ -178,7 +178,7 @@ private:
     int64_t mScanMeasureStart;
     uint32_t mScanFrameMark;   // mDeliveredFrames at measurement start
     float mScanBestFps;
-    uint8_t mScanBestMhz;
+    float mScanBestMhz;
     void finishScan();
 };
